Adjustment of accuracy

+++ assets/shaders/shader.frag
@@ -1,7 +1,7 @@
 //--- hatsuyuki ---
 // by Catzpaw 2016
 #ifdef GL_ES
-precision highp float;
+precision mediump float;
 #endif
 
 #extension GL_OES_standard_derivatives : enable
@@ -17,13 +17,13 @@ float snow(vec2 uv,float scale)
 	uv*=scale;vec2 s=floor(uv),f=fract(uv),p;float k=3.,d;
 	p=.5+.35*sin(11.*fract(sin((s+p+scale)*mat2(7,3,6,5))*5.))-f;d=length(p);k=min(d,k);
 	k=smoothstep(0.,k,sin(f.x+f.y)*0.01);
-    	return k*w;
+  return k*w;
 }
 
 void main(void){
 	vec2 uv=(gl_FragCoord.xy*2.-resolution.xy)/min(resolution.x,resolution.y); 
 	vec3 finalColor=vec3(0);
-	float c=smoothstep(1.,0.3,clamp(uv.y*.3+.8,0.,.75));
+	highp float c=smoothstep(1.,0.3,clamp(uv.y*.3+.8,0.,.75));
 	c+=snow(uv,30.)*.3;
 	c+=snow(uv,20.)*.5;
 	c+=snow(uv,15.)*.8;
